chore: extract layout from dashboard index (move-only)
